requirements.txt
Flask==1.1.2
Flask-Caching==1.8.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
dash==1.11.0
dash_bootstrap_components==0.10.0
plotly==4.6.0
pandas==1.0.3
python-dotenv==0.13.0
requests==2.23.0